app/services/api/pyproject.toml
[project]
name = "finquest-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jason Wang", email = "40612523+jasonfyw@users.noreply.github.com" }
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.118.2",
    "uvicorn[standard]>=0.32.1",
    "pydantic[email]>=2.10.3",
    "pydantic-settings>=2.7.0",
    "sqlalchemy>=2.0.34",
    "psycopg[binary]>=3.2.2",
    "supabase>=2.11.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "yfinance>=0.2.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[project.scripts]
finquest-api = "finquest_api:main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
